Update 2.3
A somewhat scoreboard has been added.

An event when triggering the powerUp will showcase a text describing the
powerUp and how to start using it. (The powerUp is active as soon as the
character hits it, but will need a new directional push (LeftArrow,
RightArrow, UpArrow, DownArrow) to start the effect. Last 6 seconds,
both the powerup effect and the text.