Textures + Lifebar
Most of the final textures has been added to the game.
A life system has been implemented which count down lives when colliding
with enemy, but doesn't do anything else as of yet.
Made some prefabs for the coins, walls and different enemies.

Things to be done:
Try to fix an issue with the power-up where each prefab generates their
own UI on top of each other.
Player position should be reset upon death and the game should end after
final death.